+++ src/lib.rs
@@ -239,5 +239,8 @@ pub mod prelude {
         Transit, TransitCreateKeyRequest, TransitKeyInfo, TransitKeyType,
     };
     #[cfg(feature = "sys")]
-    pub use crate::sys::{Capability, CapabilityView, Health, LeaderStatus, Sys};
+    pub use crate::sys::{
+        Capability, CapabilityView, Health, LeaderStatus, LoggerLevel, LoggerLevels, Sys,
+        VersionHistory, VersionHistoryEntry,
+    };
 }